More fixes -- norm and xnrm2 tests now pass on 64-bit
git-svn-id: https://svn.code.sf.net/p/freemat/code/branches/FreeMat4@3684 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ libs/libFreeMat/LAPACK.hpp
@@ -217,11 +217,11 @@ extern "C" {
 
   int xerbla_(char *srname, int *info);
 
-  float snrm2_(int *N, float *X, int *INCX);
+  double snrm2_(int *N, float *X, int *INCX);
   
   double dnrm2_(int *N, double *X, int *INCX);
 
-  float scnrm2_(int *N, float *X, int *INCX);
+  double scnrm2_(int *N, float *X, int *INCX);
   
   double dznrm2_(int *N, double *X, int *INCX);
